# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carpool Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/vehicles/user": {
            "post": {
                "tags": [
                    "User Vehicle"
                ],
                "summary": "create user vehicle",
                "description": "create user vehicle",
                "operationId": "createUserVehicle",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_id": {
                                        "description": "vehicle_id",
                                        "example": "1"
                                    },
                                    "plate_number": {
                                        "description": "plate_number",
                                        "example": "12k123r"
                                    },
                                    "color": {
                                        "description": "color",
                                        "example": "black"
                                    },
                                    "year_model": {
                                        "description": "year_model",
                                        "example": "2023"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/vehicles/user/{userVehicleId}": {
            "put": {
                "tags": [
                    "User Vehicle"
                ],
                "summary": "update user vehicle",
                "description": "update user vehicle",
                "operationId": "updateUserVehicle",
                "parameters": [
                    {
                        "name": "userVehicleId",
                        "in": "path",
                        "description": "user vehicle id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_id": {
                                        "description": "vehicle_id",
                                        "example": "1"
                                    },
                                    "plate_number": {
                                        "description": "plate_number",
                                        "example": "12k123r"
                                    },
                                    "color": {
                                        "description": "color",
                                        "example": "black"
                                    },
                                    "year_model": {
                                        "description": "year_model",
                                        "example": "2023"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Vehicle"
                ],
                "summary": "delete user vehicle",
                "description": "delete user vehicle",
                "operationId": "deleteUserVehicle",
                "parameters": [
                    {
                        "name": "userVehicleId",
                        "in": "path",
                        "description": "user vehicle id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/vehicles/user/{userVehicleId}/upload-file": {
            "post": {
                "tags": [
                    "User Vehicle"
                ],
                "summary": "upload user vehicle file",
                "description": "upload user vehicle file",
                "operationId": "uploadVehicleFile",
                "parameters": [
                    {
                        "name": "userVehicleId",
                        "in": "path",
                        "description": "user vehicle id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image",
                                        "type": "file",
                                        "example": "test"
                                    },
                                    "type": {
                                        "description": "type",
                                        "example": "vehicle_card_back,vehicle_card_front"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "",
                "name": "Authorization",
                "in": "header",
                "description": "Bearer token for Sanctum authentication"
            }
        }
    ]
}